# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Partybank Organizer Service.",
        "title": "Partybank Organizer Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://www.thepartybank.com",
            "email": "unavailable"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "partybank-organizer-269c8057a65f.herokuapp.com",
    "paths": {
...
        "/api/v1/event/staff": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Adds Event Staff",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organizer"
                ],
                "summary": "Add Event Staff",
                "parameters": [
                    {
                        "description": "Organizer tags",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.AddEventStaffRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/series": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Adds New Series",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Series"
                ],
                "summary": "Add New Series",
                "parameters": [
                    {
                        "description": "Series tags",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateSeriesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/series/organizer/{organizerId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get Series by organizerId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Series"
                ],
                "summary": "Get Series by organizerId",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "organizerId",
                        "name": "organizerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/series/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get Series by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Series"
                ],
                "summary": "Get Series by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "series id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ticket": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get Ticket By id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get Ticket By id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ticketId",
                        "name": "ticketId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.TicketResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Add Ticket to Event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Add Ticket to Event",
                "parameters": [
                    {
                        "description": "Ticket tags",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.TicketResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ticket/update": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update Ticket By reference",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Update Ticket By reference",
                "parameters": [
                    {
                        "type": "string",
                        "description": "reference",
                        "name": "reference",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.TicketResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ticket/{eventId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all Tickets for Event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get all Tickets for Event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "eventId",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Ticket"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.AddEventStaffRequest": {
            "type": "object",
            "required": [
                "event_id"
            ],
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "staff_emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dtos.CreateSeriesRequest": {
            "type": "object",
            "required": [
                "organizer_id"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organizer_id": {
                    "type": "integer"
                },
                "series_logo": {
                    "type": "string"
                }
            }
        },
        "dtos.CreateTicketRequest": {
            "type": "object",
            "required": [
                "event_id"
            ],
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "colour": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "is_transfer_payment_fees_to_guest": {
                    "description": "TODO: Default: false",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "price_change_date": {
                    "type": "string"
                },
                "price_change_time": {
                    "type": "string"
                },
                "purchase_limit": {
                    "type": "integer"
                },
                "stock": {
                    "type": "string"
                },
                "ticket_perks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ticket_sale_end_date": {
                    "type": "string"
                },
                "ticket_sale_start_date": {
                    "type": "string"
                },
                "ticket_sale_start_time": {
                    "type": "string"
                },
                "ticket_sales_end_time": {
                    "type": "string"
                },
                "ticket_type": {
                    "type": "string"
                }
            }
        },
...
        "dtos.RaveResponse": {
            "type": "object"
        },
        "dtos.TicketResponse": {
            "type": "object",
            "properties": {
                "additional_information_fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "available_discounted_tickets": {
                    "type": "integer"
                },
                "capacity": {
                    "type": "integer"
                },
                "colour": {
                    "type": "string"
                },
                "discount_code": {
                    "type": "string"
                },
                "discount_price": {
                    "type": "number"
                },
                "discount_type": {
                    "type": "string"
                },
                "is_ticket_sale_date_expired": {
                    "type": "boolean"
                },
                "is_transfer_payment_fees_to_guest": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "number_in_stock": {
                    "type": "integer"
                },
                "percentage": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "purchase_limit": {
                    "type": "integer"
                },
                "stock": {
                    "type": "string"
                },
                "ticket_perks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ticket_reference": {
                    "type": "string"
                },
                "ticket_sale_end_date": {
                    "type": "string"
                },
                "ticket_sale_start_date": {
                    "type": "string"
                },
                "ticket_sale_start_time": {
                    "type": "string"
                },
                "ticket_sales_end_time": {
                    "type": "string"
                },
                "ticket_type": {
                    "type": "string"
                }
            }
        },
...
        "models.ActivePeriod": {
            "type": "object",
            "properties": {
                "ticket_sale_end_date": {
                    "type": "string"
                },
                "ticket_sale_end_time": {
                    "type": "string"
                },
                "ticket_sale_start_date": {
                    "type": "string"
                },
                "ticket_sale_start_time": {
                    "type": "string"
                }
            }
        },
...
        "models.Ticket": {
            "type": "object",
            "properties": {
                "active_period": {
                    "$ref": "#/definitions/models.ActivePeriod"
                },
                "additional_information_fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "available_discounted_tickets": {
                    "type": "integer"
                },
                "capacity": {
                    "type": "integer"
                },
                "category": {
                    "type": "integer"
                },
                "colour": {
                    "type": "string"
                },
                "discount_code": {
                    "type": "string"
                },
                "discount_price": {
                    "type": "number"
                },
                "discount_type": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "event_reference": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_sold_out_ticket": {
                    "type": "boolean"
                },
                "is_transfer_payment_fees_to_guest": {
                    "type": "boolean"
                },
                "max_seats": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "number_available": {
                    "type": "integer"
                },
                "percentage": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "purchase_limit": {
                    "type": "integer"
                },
                "reference": {
                    "type": "string"
                },
                "reserved": {
                    "type": "integer"
                },
                "stock": {
                    "type": "string"
                },
                "ticket_perks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}